refactor(svm-utils): split utility functions into separate files (#612)
* WIP


* WIP


* WIP


* WIP


* WIP


* WIP


* WIP


* WIP


* refactor(svm-utils): split utility functions into separate files

* fix: merge


* fix: merge


* feat: delete utils


---------
